Added docstrings. Moved nautobot_app to app. Added PermissionsMixin. Added isinstance and issubclass checking. Added new choiceset for button colors.
