+++ nautobot/extras/registry.py
@@ -23,8 +23,6 @@ class Registry(dict):
 
 
 registry = Registry()
-registry["nav_menu"] = {"tabs": {}}
-
 
 DatasourceContent = namedtuple("DatasourceContent", ["name", "content_identifier", "icon", "callback"])
 """
